+++ src/views/GovernanceReviewTeam/Actions/RejectIntake.tsx
@@ -12,7 +12,7 @@ import FieldGroup from 'components/shared/FieldGroup';
 import HelpText from 'components/shared/HelpText';
 import Label from 'components/shared/Label';
 import TextAreaField from 'components/shared/TextAreaField';
-import useSystemIntake from 'hooks/useSystemIntake';
+import useSystemIntakeContacts from 'hooks/useSystemIntakeContacts';
 import RejectIntakeQuery from 'queries/RejectIntakeQuery';
 import {
   RejectIntake as RejectIntakeType,
@@ -28,7 +28,6 @@ import EmailRecipientsFields from './EmailRecipientsFields';
 
 const RejectIntake = () => {
   const { systemId } = useParams<{ systemId: string }>();
-  const { systemIntake } = useSystemIntake(systemId);
   const history = useHistory();
   const { t } = useTranslation('action');
   const [shouldSendEmail, setShouldSendEmail] = useState<boolean>(true);
@@ -40,6 +39,15 @@ const RejectIntake = () => {
     errorPolicy: 'all'
   });
 
+  // Requester object and loading state
+  const {
+    contacts: {
+      data: { requester },
+      loading
+    }
+  } = useSystemIntakeContacts(systemId);
+
+  // Active contact for adding/verifying recipients
   const [
     activeContact,
     setActiveContact
@@ -52,7 +60,7 @@ const RejectIntake = () => {
     nextSteps: '',
     reason: '',
     notificationRecipients: {
-      regularRecipientEmails: [systemIntake?.requester?.email!],
+      regularRecipientEmails: requester.id ? [requester.email] : [],
       shouldNotifyITGovernance: true,
       shouldNotifyITInvestment: false
     }
@@ -78,6 +86,9 @@ const RejectIntake = () => {
       }
     });
   };
+
+  // Wait for contacts to load before returning form
+  if (loading) return null;
 
   return (
     <Formik
