+++ src/views/GovernanceReviewTeam/Actions/SubmitAction.test.tsx
@@ -6,8 +6,8 @@ import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import configureMockStore from 'redux-mock-store';
 
-import grtActions from 'constants/grtActions';
 import { businessCaseInitialData } from 'data/businessCase';
+import { contactQueries, grtActions, mockData } from 'data/mock/grtActions';
 import { initialSystemIntakeForm } from 'data/systemIntake';
 import CreateSystemIntakeActionBusinessCaseNeeded from 'queries/CreateSystemIntakeActionBusinessCaseNeededQuery';
 import CreateSystemIntakeActionBusinessCaseNeedsChanges from 'queries/CreateSystemIntakeActionBusinessCaseNeedsChangesQuery';
@@ -18,7 +18,6 @@ import CreateSystemIntakeActionNotRespondingClose from 'queries/CreateSystemInta
 import CreateSystemIntakeActionReadyForGRT from 'queries/CreateSystemIntakeActionReadyForGRTQuery';
 import CreateSystemIntakeActionSendEmail from 'queries/CreateSystemIntakeActionSendEmailQuery';
 import GetSystemIntakeQuery from 'queries/GetSystemIntakeQuery';
-import { GetSystemIntakeContactsQuery } from 'queries/SystemIntakeContactsQueries';
 
 import RequestOverview from '../RequestOverview';
 
@@ -28,118 +27,15 @@ const waitForPageLoad = async (testId: string = 'grt-submit-action-view') => {
   });
 };
 
-const systemIntakeId = 'a4158ad8-1236-4a55-9ad5-7e15a5d49de2';
+// Mock system intake and contacts data
+const {
+  systemIntake,
+  contacts: { requester, productManager }
+} = mockData;
+const systemIntakeId = systemIntake.id;
 
-const requester = {
-  euaUserId: 'SF13',
-  name: 'Jerry Seinfeld',
-  component: 'Center for Medicaid and CHIP Services',
-  email: 'sf13@local.fake'
-};
-
-const businessOwner = {
-  systemIntakeId,
-  role: 'Business Owner',
-  euaUserId: 'SF13',
-  commonName: 'Jerry Seinfeld',
-  component: 'Center for Medicaid and CHIP Services',
-  email: 'sf13@local.fake',
-  id: 'b25bd13b-72a0-4d0e-b5be-7852a1a8259d'
-};
-
-const productManager = {
-  systemIntakeId,
-  euaUserId: 'ABCD',
-  role: 'Product Manager',
-  commonName: 'Jane Doe',
-  component: 'Center for Program Integrity',
-  email: 'abcd@local.fake',
-  id: '' // Leave out ID so contact shows as unverified
-};
-
-const isso = {
-  systemIntakeId,
-  euaUserId: 'WXYZ',
-  role: 'ISSO',
-  commonName: 'John Smith',
-  component: 'CMS Wide',
-  email: 'wxyz@local.fake',
-  id: '346eefa9-c635-4c0b-bc29-26f272c33d0c'
-};
-
-const systemIntake = {
-  id: systemIntakeId,
-  euaUserId: requester.euaUserId,
-  adminLead: '',
-  businessNeed: 'Test business need',
-  businessSolution: 'Test business solution',
-  businessOwner: {
-    component: businessOwner.component,
-    name: businessOwner.commonName
-  },
-  contract: {
-    contractor: 'Contractor Name',
-    endDate: {
-      day: '31',
-      month: '12',
-      year: '2023'
-    },
-    hasContract: 'HAVE_CONTRACT',
-    startDate: {
-      day: '1',
-      month: '1',
-      year: '2021'
-    },
-    vehicle: 'Sole source',
-    number: '123456-7890'
-  },
-  costs: {
-    isExpectingIncrease: 'YES',
-    expectedIncreaseAmount: '10 million dollars'
-  },
-  currentStage: 'I have done some initial research',
-  decisionNextSteps: '',
-  grbDate: null,
-  grtDate: null,
-  grtFeedbacks: [],
-  governanceTeams: {
-    isPresent: true,
-    teams: []
-  },
-  isso: {
-    isPresent: true,
-    component: isso.component,
-    name: isso.commonName
-  },
-  existingFunding: true,
-  fundingSources: [{ fundingNumber: '123456', source: 'Research' }],
-  lcid: '',
-  lcidExpiresAt: null,
-  lcidScope: '',
-  lcidCostBaseline: null,
-  needsEaSupport: true,
-  productManager: {
-    component: productManager.component,
-    name: productManager.commonName
-  },
-  rejectionReason: '',
-  requester: {
-    component: requester.component,
-    email: requester.email,
-    name: requester.name
-  },
-  requestName: 'TACO',
-  requestType: 'NEW',
-  status: 'INTAKE_SUBMITTED',
-  createdAt: null,
-  submittedAt: null,
-  updatedAt: null,
-  archivedAt: null,
-  decidedAt: null,
-  businessCaseId: null,
-  lastAdminNote: null,
-  grtReviewEmailBody: ''
-};
+// Mock contact queries
+const { getSystemIntakeContactsQuery, getCedarContactsQuery } = contactQueries;
 
 jest.mock('@okta/okta-react', () => ({
   useOktaAuth: () => {
@@ -175,22 +71,6 @@ describe('Submit Action', () => {
     }
   };
 
-  const contactsQuery = {
-    request: {
-      query: GetSystemIntakeContactsQuery,
-      variables: {
-        id: systemIntakeId
-      }
-    },
-    result: {
-      data: {
-        systemIntakeContacts: {
-          systemIntakeContacts: [businessOwner, productManager, isso]
-        }
-      }
-    }
-  };
-
   const mockStore = configureMockStore();
   const defaultStore = mockStore({
     auth: {
@@ -213,7 +93,7 @@ describe('Submit Action', () => {
   });
 
   const renderActionPage = (slug: string, mocks: any[]) => {
-    render(
+    return render(
       <MemoryRouter
         initialEntries={[
           `/governance-review-team/${systemIntakeId}/actions/${slug}`
@@ -274,30 +154,12 @@ describe('Submit Action', () => {
   );
 
   describe('actions', () => {
-    it.only('renders additional contacts', async () => {
-      const { asFragment } = render(
-        <MemoryRouter
-          initialEntries={[
-            `/governance-review-team/${systemIntakeId}/actions/not-it-request`
-          ]}
-        >
-          <MockedProvider
-            mocks={[intakeQuery, contactsQuery]}
-            addTypename={false}
-          >
-            <Provider store={defaultStore}>
-              <Route
-                path={[
-                  '/governance-review-team/:systemId/intake-request',
-                  `/governance-review-team/:systemId/actions/not-it-request`
-                ]}
-              >
-                <RequestOverview />
-              </Route>
-            </Provider>
-          </MockedProvider>
-        </MemoryRouter>
-      );
+    it('renders additional contacts', async () => {
+      const { asFragment } = renderActionPage('not-it-request', [
+        intakeQuery,
+        getCedarContactsQuery,
+        getSystemIntakeContactsQuery
+      ]);
 
       await waitForPageLoad();
       screen.getByTestId('truncatedContentButton').click();
@@ -309,7 +171,7 @@ describe('Submit Action', () => {
 
       screen.getByTestId('button_verify-recipient').click();
       expect(screen.getByRole('combobox', { name: 'Cedar-Users' })).toHaveValue(
-        `${productManager.commonName}, ${productManager.euaUserId}`
+        productManager.commonName
       );
 
       expect(asFragment()).toMatchSnapshot();
@@ -344,7 +206,11 @@ describe('Submit Action', () => {
         }
       };
 
-      renderActionPage('not-it-request', [intakeQuery, notITRequestMutation]);
+      renderActionPage('not-it-request', [
+        intakeQuery,
+        getSystemIntakeContactsQuery,
+        notITRequestMutation
+      ]);
       await waitForPageLoad();
 
       expect(
@@ -354,7 +220,7 @@ describe('Submit Action', () => {
       // Check requester is set as default recipient
       expect(
         screen.getByRole('checkbox', {
-          name: `${requester.name}, ${requester.component} (Requester)`
+          name: `${requester.commonName}, ${requester.component} (Requester)`
         })
       ).toBeChecked();
 
@@ -442,7 +308,11 @@ describe('Submit Action', () => {
         }
       };
 
-      renderActionPage('need-biz-case', [intakeQuery, needBizCaseMutation]);
+      renderActionPage('need-biz-case', [
+        intakeQuery,
+        getSystemIntakeContactsQuery,
+        needBizCaseMutation
+      ]);
       await waitForPageLoad();
 
       expect(
@@ -452,7 +322,7 @@ describe('Submit Action', () => {
       // Check requester is set as default recipient
       expect(
         screen.getByRole('checkbox', {
-          name: `${requester.name}, ${requester.component} (Requester)`
+          name: `${requester.commonName}, ${requester.component} (Requester)`
         })
       ).toBeChecked();
 
@@ -499,7 +369,11 @@ describe('Submit Action', () => {
         }
       };
 
-      renderActionPage('ready-for-grt', [intakeQuery, readyForGRTMutation]);
+      renderActionPage('ready-for-grt', [
+        intakeQuery,
+        getSystemIntakeContactsQuery,
+        readyForGRTMutation
+      ]);
       await waitForPageLoad();
 
       expect(screen.getByText(/Mark as ready for GRT/i)).toBeInTheDocument();
@@ -507,7 +381,7 @@ describe('Submit Action', () => {
       // Check requester is set as default recipient
       expect(
         screen.getByRole('checkbox', {
-          name: `${requester.name}, ${requester.component} (Requester)`
+          name: `${requester.commonName}, ${requester.component} (Requester)`
         })
       ).toBeChecked();
 
@@ -556,6 +430,7 @@ describe('Submit Action', () => {
 
       renderActionPage('biz-case-needs-changes', [
         intakeQuery,
+        getSystemIntakeContactsQuery,
         needsChangesMutation
       ]);
       await waitForPageLoad();
@@ -569,7 +444,7 @@ describe('Submit Action', () => {
       // Check requester is set as default recipient
       expect(
         screen.getByRole('checkbox', {
-          name: `${requester.name}, ${requester.component} (Requester)`
+          name: `${requester.commonName}, ${requester.component} (Requester)`
         })
       ).toBeChecked();
 
@@ -616,7 +491,11 @@ describe('Submit Action', () => {
         }
       };
 
-      renderActionPage('no-governance', [intakeQuery, noGovernanceMutation]);
+      renderActionPage('no-governance', [
+        intakeQuery,
+        getSystemIntakeContactsQuery,
+        noGovernanceMutation
+      ]);
       await waitForPageLoad();
 
       expect(screen.getByText(/Close project/i)).toBeInTheDocument();
@@ -624,7 +503,7 @@ describe('Submit Action', () => {
       // Check requester is set as default recipient
       expect(
         screen.getByRole('checkbox', {
-          name: `${requester.name}, ${requester.component} (Requester)`
+          name: `${requester.commonName}, ${requester.component} (Requester)`
         })
       ).toBeChecked();
 
@@ -677,13 +556,17 @@ describe('Submit Action', () => {
         }
       };
 
-      renderActionPage('send-email', [intakeQuery, sendEmailMutation]);
+      renderActionPage('send-email', [
+        intakeQuery,
+        getSystemIntakeContactsQuery,
+        sendEmailMutation
+      ]);
       await waitForPageLoad();
 
       // Check requester is set as default recipient
       expect(
         screen.getByRole('checkbox', {
-          name: `${requester.name}, ${requester.component} (Requester)`
+          name: `${requester.commonName}, ${requester.component} (Requester)`
         })
       ).toBeChecked();
 
@@ -732,6 +615,7 @@ describe('Submit Action', () => {
 
       renderActionPage('guide-received-close', [
         intakeQuery,
+        getSystemIntakeContactsQuery,
         guideReceivedCloseMutation
       ]);
       await waitForPageLoad();
@@ -743,7 +627,7 @@ describe('Submit Action', () => {
       // Check requester is set as default recipient
       expect(
         screen.getByRole('checkbox', {
-          name: `${requester.name}, ${requester.component} (Requester)`
+          name: `${requester.commonName}, ${requester.component} (Requester)`
         })
       ).toBeChecked();
 
@@ -792,6 +676,7 @@ describe('Submit Action', () => {
 
       renderActionPage('not-responding-close', [
         intakeQuery,
+        getSystemIntakeContactsQuery,
         notRespondingCloseMutation
       ]);
       await waitForPageLoad();
@@ -803,7 +688,7 @@ describe('Submit Action', () => {
       // Check requester is set as default recipient
       expect(
         screen.getByRole('checkbox', {
-          name: `${requester.name}, ${requester.component} (Requester)`
+          name: `${requester.commonName}, ${requester.component} (Requester)`
         })
       ).toBeChecked();
